+++ example/lib/widgets/Modifications.dart
@@ -49,7 +49,7 @@ class _ModificationsState extends State<Modifications> {
 
     myFlag = currentVisitor?.getFlag(keyFlagController.text, defaultValue);
 
-    var ret = myFlag?.value(vistorExposed: false);
+    var ret = myFlag?.value(visitorExposed: false);
 
     //  var ret =
     //   currentVisitor?.getModification(keyFlagController.text, defaultValue);